trying to fix

+++ Jenkinsfile
@@ -25,8 +25,14 @@ node('host_ci') {
    }
    stage('Docker run') {
       //Deploying the docker image, meaning: runnig the host-status web application
-      sh "docker rm -f `docker ps -a -f name=host-status-web -q`"
+      def ps_list = sh(script: 'docker ps -a -f name=host-status-web -q | wc -l', returnStdout: true)
+      if ( ps_list > 0 ) {
+          sh "docker rm -f `docker ps -a -f name=host-status-web -q`"
+      }
       sh "docker run -d --name host-status-web -p 8085:8085 host-status-web"
-      sh "docker rmi `docker images -f dangling=true -q`"
+      def image_list = sh(script: 'docker images -f dangling=true -q | wc -l', returnStdout: true)
+      if ( image_list > 0 )  {
+          sh "docker rmi `docker images -f dangling=true -q`"
+      }
    }
 }
